ProductController: long parameter type changed to int type.
GetProductById parameter type changed to int from long.

attached() added in product-list.ts.

In progress:

href towards localhost:aurelia/product/:id

api-endpoint edited by adding a slash (/) at end of URL.

+++ Webshop.Data/Services/ProductService.cs
@@ -23,7 +23,7 @@
             return await DbContext.Products.Where(i => i.Id != 0).ToListAsync();
         }
 
-        public async Task<Product> GetProductById(long id)
+        public async Task<Product> GetProductById(int id)
         {
             return await DbContext.Products.FindAsync(id);
         }
